fix: convo resets on model change
